+++ packages/db/src/db.ts
@@ -73,10 +73,11 @@ export interface CollectionRules<
   // update?: Rule<M>[];
 }
 
-interface TransactOptions {
+export interface TransactOptions {
   storeScope?: { read: string[]; write: string[] };
   skipRules?: boolean;
   dangerouslyBypassSchemaInitialization?: boolean;
+  manualSchemaRefresh?: boolean;
 }
 
 export type CreateCollectionOperation = [
@@ -739,6 +740,7 @@ export default class DB<M extends Models<any, any> | undefined = undefined> {
           schema,
           skipRules: options.skipRules,
           logger: this.logger.scope('tx'),
+          manualSchemaRefresh: options.manualSchemaRefresh,
         });
         return await callback(tx);
       }, options.storeScope);
@@ -919,14 +921,14 @@ export default class DB<M extends Models<any, any> | undefined = undefined> {
       return unsub;
     };
 
-    const unsubPromise = startSubscription();
+    const unsubPromise = startSubscription().catch(onError);
 
     return async () => {
       // Immediately set unsubscribed to true to prevent any new results from being processed
       unsubscribed = true;
       this.logger.debug('subscribe END', { query });
       const unsub = await unsubPromise;
-      return unsub();
+      return unsub?.();
     };
   }
 
@@ -970,11 +972,11 @@ export default class DB<M extends Models<any, any> | undefined = undefined> {
       return unsub;
     };
 
-    const unsubPromise = startSubscription();
+    const unsubPromise = startSubscription().catch(onError);
 
     return async () => {
       const unsub = await unsubPromise;
-      return unsub();
+      return unsub?.();
     };
   }