+++ packages/angular/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @triplit/angular
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [40e2e50]
+  - @triplit/client@0.3.80
+
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [1c0bdf7]
+  - @triplit/client@0.3.79
+
 ## 0.0.4
 
 ### Patch Changes