+++ packages/react/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @triplit/react
 
+## 0.3.84
+
+### Patch Changes
+
+- Updated dependencies [40e2e50]
+  - @triplit/client@0.3.80
+
+## 0.3.83
+
+### Patch Changes
+
+- Updated dependencies [1c0bdf7]
+  - @triplit/client@0.3.79
+
 ## 0.3.82
 
 ### Patch Changes