+++ packages/server-core/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @triplit/server-core
 
+## 0.3.65
+
+### Patch Changes
+
+- Updated dependencies [40e2e50]
+  - @triplit/db@0.3.62
+
+## 0.3.64
+
+### Patch Changes
+
+- Updated dependencies [c78aa65]
+- Updated dependencies [6ccbbd5]
+  - @triplit/db@0.3.61
+
 ## 0.3.63
 
 ### Patch Changes