+++ packages/server/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @triplit/server
 
+## 0.3.70
+
+### Patch Changes
+
+- Updated dependencies [40e2e50]
+  - @triplit/db@0.3.62
+  - @triplit/server-core@0.3.65
+
+## 0.3.69
+
+### Patch Changes
+
+- Updated dependencies [c78aa65]
+- Updated dependencies [6ccbbd5]
+  - @triplit/db@0.3.61
+  - @triplit/server-core@0.3.64
+
 ## 0.3.68
 
 ### Patch Changes